backend/visual_ai_server/requirements.txt
Package                      Version
---------------------------- ---------
absl-py                      2.2.1
annotated-types              0.7.0
anyio                        4.9.0
astunparse                   1.6.3
certifi                      2025.1.31
charset-normalizer           3.4.1
click                        8.1.8
colorama                     0.4.6
exceptiongroup               1.2.2
fastapi                      0.109.2
filelock                     3.18.0
flatbuffers                  25.2.10
fsspec                       2025.3.2
gast                         0.6.0
google-pasta                 0.2.0
grpcio                       1.71.0
h11                          0.14.0
h5py                         3.13.0
idna                         3.10
importlib_metadata           8.6.1
Jinja2                       3.1.6
keras                        3.9.0
libclang                     18.1.1
Markdown                     3.7
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
mdurl                        0.1.2
ml_dtypes                    0.5.1
mpmath                       1.3.0
namex                        0.0.8
networkx                     3.2.1
numpy                        1.26.4
opencv-python-headless       4.9.0.80
opt_einsum                   3.4.0
optree                       0.14.1
packaging                    24.2
pillow                       10.2.0
pip                          25.1.1
protobuf                     5.29.4
pydantic                     2.6.1
pydantic_core                2.27.2
Pygments                     2.19.1
python-multipart             0.0.9
RapidFuzz                    3.12.2
requests                     2.32.3
rich                         13.9.4
setuptools                   58.1.0
six                          1.17.0
sniffio                      1.3.1
starlette                    0.46.1
sympy                        1.13.1
tensorboard                  2.19.0
tensorboard-data-server      0.7.2
tensorflow                   2.19.0
tensorflow-hub               0.16.1
tensorflow-io-gcs-filesystem 0.31.0
termcolor                    2.5.0
tf_keras                     2.19.0
torch                        2.6.0
torchaudio                   2.6.0
torchvision                  0.21.0
typing_extensions            4.13.2
typing-inspection            0.4.0
urllib3                      2.3.0
uvicorn                      0.27.1
Werkzeug                     3.1.3
wheel                        0.45.1
wrapt                        1.17.2
zipp                         3.21.0
aiofiles                     23.2.1
sqlalchemy                   2.0.27
pybind11                     2.13.6
pytest                       8.0.0
httpx                        0.27.0
python-jose                  3.3.0
passlib                      1.7.4
bcrypt                       4.1.2